requirements.txt
streamlit
requests
tifffile
pandas
numpy
matplotlib.pyplot
io
